Changed name of program to effect zone
also began working on setting up multiple plots
