+++ .idea/compiler.xml
@@ -6,11 +6,11 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
-        <module name="PlayAreaPlugin" />
+        <module name="EffectZonePlugin" />
       </profile>
     </annotationProcessing>
     <bytecodeTargetLevel>
-      <module name="PlayAreaPlugin" target="1.8" />
+      <module name="EffectZonePlugin" target="1.8" />
     </bytecodeTargetLevel>
   </component>
 </project>
