+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/PlayAreaPlugin.iml" filepath="$PROJECT_DIR$/.idea/PlayAreaPlugin.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/EffectZonePlugin.iml" filepath="$PROJECT_DIR$/.idea/EffectZonePlugin.iml" />
     </modules>
   </component>
 </project>
