+++ README.txt
@@ -1,4 +1,4 @@
-# PlayAreaPlugin
+# EffectZonePlugin
 Use a tool to create a square play area that players roam in.
 If a player wanders out of this square, a customizable message will be sent, as well as customizable potion effects, & customizable mob damage
