+++ pom.xml
@@ -5,11 +5,11 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>me.Jalawaquin</groupId>
-    <artifactId>PlayArea</artifactId>
+    <artifactId>EffectZone</artifactId>
     <version>1.0-SNAPSHOT</version>
     <packaging>jar</packaging>
 
-    <name>PlayArea</name>
+    <name>EffectZone</name>
 
     <properties>
         <java.version>1.8</java.version>
