+++ src/main/java/me/jalawaquin/effectzone/commands/setEffectZone.java
@@ -1,0 +1,36 @@
+package me.jalawaquin.effectzone.commands;
+
+import org.bukkit.ChatColor;
+import org.bukkit.Material;
+import org.bukkit.command.Command;
+import org.bukkit.command.CommandExecutor;
+import org.bukkit.command.CommandSender;
+import org.bukkit.entity.Player;
+import org.bukkit.inventory.ItemStack;
+import org.bukkit.inventory.meta.ItemMeta;
+
+public class setEffectZone implements CommandExecutor {
+    @Override
+    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
+        if (sender instanceof Player) {
+            Player player = (Player) sender;
+
+            if(player.hasPermission("playarea.seteffectzone")){
+                ItemStack itemstack = new ItemStack(Material.GOLDEN_HOE, 1);
+                ItemMeta meta = itemstack.getItemMeta();
+
+                player.sendMessage("Left Click a corner of your desired effect zone");
+
+                meta.setDisplayName(ChatColor.LIGHT_PURPLE + "Effect Zone Tool");
+
+                itemstack.setItemMeta(meta);
+                player.getInventory().addItem(itemstack);
+            }
+            else{
+                player.sendMessage(ChatColor.RED + "You do not have the required permissions to use this command.");
+            }
+        }
+
+        return false;
+    }
+}
