--- src/main/java/me/jalawaquin/playarea/PlayArea.java
@@ -1,31 +1,0 @@
-package me.jalawaquin.playarea;
-
-import me.jalawaquin.playarea.commands.deletePlayArea;
-import me.jalawaquin.playarea.commands.insidePlayArea;
-import me.jalawaquin.playarea.commands.outsidePlayArea;
-import me.jalawaquin.playarea.commands.turnPlayArea;
-import me.jalawaquin.playarea.commands.setPlayArea;
-
-import me.jalawaquin.playarea.events.PlayEvents;
-import me.jalawaquin.playarea.listeners.PlotAreaListener;
-import me.jalawaquin.playarea.settings.Plots;
-import org.bukkit.plugin.java.JavaPlugin;
-
-public final class PlayArea extends JavaPlugin {
-    //turn into ArrayList to have multiple plots ? (if I do this need add plot and remove plot functionality etc.)
-    private Plots plot = new Plots();
-    @Override
-    public void onEnable() {
-        // Plugin startup logic
-
-        getCommand("deletePlayArea").setExecutor(new deletePlayArea(plot));
-        getCommand("insidePlayArea").setExecutor(new insidePlayArea(plot));
-        getCommand("outsidePlayArea").setExecutor(new outsidePlayArea(plot));
-        getCommand("setPlayArea").setExecutor(new setPlayArea());
-        getCommand("turnPlayArea").setExecutor(new turnPlayArea(plot));
-
-        getServer().getPluginManager().registerEvents(new PlayEvents(plot),this);
-        getServer().getPluginManager().registerEvents(new PlotAreaListener(plot), this);
-    }
-
-}
