+++ src/main/resources/plugin.yml
@@ -1,21 +1,21 @@
-name: PlayArea
+name: EffectZone
 version: ${project.version}
-main: me.jalawaquin.playarea.PlayArea
+main: me.jalawaquin.effectzone.EffectZone
 api-version: 1.13
 authors: [Jalawaquin]
 commands:
-  setPlayArea:
-    description: set the play area
-    aliases: setplayarea
-  deletePlayArea:
-    description: delete the play area
-    aliases: deleteplayarea
-  turnPlayArea:
-    description: turn play area modifications on and off
-    aliases: turnplayarea
-  insidePlayArea:
-    description: modify the mob or potion effects when inside play area
-    aliases: insideplayarea
+  setEffectZone:
+    description: set the effect zone
+    aliases: seteffectzone
+  deleteEffectZone:
+    description: delete the effectzone
+    aliases: deleteeffectzone
+  turnEffectZone:
+    description: turn effect zone modifications on and off
+    aliases: turneffectzone
+  insideEffectZone:
+    description: modify the mob or potion effects when inside effect zone
+    aliases: insideeffectzone
   outsidePlayArea:
-    description: modify the mob or potion effects when inside play area
-    aliases: outsideplayarea
+    description: modify the mob or potion effects when inside effect zone
+    aliases: outsideeffectzone
